lint cleanup of tools/stationxml
